Fix: payment and pricing page
